pyproject.toml
[tool.poetry]
name = "hackapizzaiug-restaurant"
version = "0.1.0"
description = ""
authors = ["alessiogandelli <alessiogandelli99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
datapizza-ai = "^0.0.9"
aiohttp = "^3.13.3"
python-dotenv = "^1.2.1"
datapizza-ai-clients-openai-like = "^0.0.9"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
